% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_table_peak.R
\name{clean_table_peak}
\alias{clean_table_peak}
\title{Clean metabolites from Table of Peaks}
\usage{
clean_table_peak(table, min_QC_RSD = NULL, min_Perc_missing = NULL)
}
\arguments{
\item{min_QC_RSD}{The Relative Standard Deviation}

\item{min_Perc_missing}{The percentage of absent/missing values}

\item{table_peak}{The data frame containing metabolite
information including QC_min_QC_RSD and min_Perc_missing.}
}
\value{
A data frame containing only clean metabolites.
}
\description{
\code{clean_table_peak} filters clean metabolites from the provided
table_peak based on criteria, such as missing values and
standard deviation of peaks.
}
\examples{
library(MDAW)
peaks <- MDAW::table_peak
clean_metabolites <- clean_table_peak(peaks)

}
