% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_binary_Y.R
\name{create_binary_Y}
\alias{create_binary_Y}
\title{Create a binary Y vector for stratifying the samples}
\usage{
create_binary_Y(dataTable, classColumn)
}
\arguments{
\item{dataTable}{A data table containing the class labels.}

\item{classColumn}{The name of the column in the data table that corresponds to the class labels.}
}
\value{
A binary vector where the class labels are converted into binary values (1 for one group, 0 for the other).
}
\description{
This function takes a data table and creates a binary Y vector based on a specified column.
The column should correspond to the class labels, with two distinct groups.
}
\examples{
dataTable <- data.frame(Class = c("GC", "HE", "GC", "HE"))
binaryY <- create_binary_Y(dataTable, "Class")
}
