% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data_train_test_customized.R
\name{custom_train_test_split}
\alias{custom_train_test_split}
\title{Custom Train-Test Split Function}
\usage{
custom_train_test_split(dataTable2, Y, test_size = 0.25, seed = NULL)
}
\arguments{
\item{dataTable2}{The dataset to be split.}

\item{Y}{List of binary outcomes for stratification.}

\item{test_size}{Proportion of the dataset to include in the test set.}

\item{seed}{Seed for random number generation.}
}
\value{
A list containing:
  \item{dataTrain}{Training dataset (dataframe)}
  \item{dataTest}{Test dataset (dataframe)}
  \item{Ytrain}{Known outcomes for the training dataset}
  \item{Ytest}{Known outcomes for the test dataset}
}
\description{
This function splits the dataset and binary outcomes into training and test sets.
It replicates the functionality of scikit-learn's train_test_split function.
}
\examples{
dataTable2 <- data.frame(matrix(rnorm(100*10), nrow=100, ncol=10))
Y <- sample(c(0,1), 100, replace = TRUE)
split_data <- custom_train_test_split(dataTable2, Y, test_size = 0.25, seed = 123)
dim(split_data$dataTrain)
dim(split_data$dataTest)
}
