% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dataXL.R
\name{import_dataXL}
\alias{import_dataXL}
\title{Import Data and Peak tables from xslx}
\usage{
import_dataXL(xlsx_file_name, sheet_name)
}
\arguments{
\item{xlsx_file_name}{A name of excel file.}

\item{sheet_name}{A name of indicating the sheet of interest.}
}
\value{
The table of the `tibble` format.
}
\description{
[import_dataXL()] reads sheets from xslx and load them.
}
\examples{

import_dataXL("data-raw/Gastric_NMR.xlsx", "data")

}
