% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{pca_interaction_plot}
\alias{pca_interaction_plot}
\title{Perform PCA and Generate PCA Plot with Interaction}
\usage{
pca_interaction_plot(data, categories)
}
\arguments{
\item{data}{A data frame containing the input data, where the variables of interest are columns 6 and onwards.}

\item{categories}{A character vector specifying the names of the categorical variables used for interaction coloring.}
}
\value{
None. The function generates the PCA plot with interaction.
}
\description{
This function performs Principal Component Analysis (PCA) on the input data matrix
and generates a PCA plot with interaction using the ggplot2 and factoextra packages.
}
\examples{
# Example usage
# pca_interaction_plot(data = table_data, categories = c("Category1", "Category2"))

}
\seealso{
\code{\link{prcomp}}, \code{\link{fviz_pca_ind}}, \code{\link{ggplot2}}, \code{\link{theme_minimal}}
}
