% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_descriptive_stats.R
\name{plot_metabolite_boxplots}
\alias{plot_metabolite_boxplots}
\title{Visualize metabolite concentrations by sample type}
\usage{
plot_metabolite_boxplots(data)
}
\arguments{
\item{data}{A data frame containing metabolomics data.}
}
\value{
A ggplot object displaying boxplots of metabolite concentrations by sample type.
}
\description{
This function creates boxplots to visualize the distribution of metabolite concentrations by sample type.
}
\examples{
data <- MDAW::table_data
plot_metabolite_boxplots(data)

}
