% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_descriptive_stats.R
\name{plot_metabolite_boxplots_by_class}
\alias{plot_metabolite_boxplots_by_class}
\title{Visualize metabolite concentrations by class}
\usage{
plot_metabolite_boxplots_by_class(data)
}
\arguments{
\item{data}{A data frame containing metabolomics data.}
}
\value{
A ggplot object displaying boxplots of metabolite concentrations by class.
}
\description{
This function creates boxplots to visualize the distribution of metabolite concentrations by class.
}
\examples{
data <- MDAW::table_data
plot_metabolite_boxplots_by_class(data)

}
