% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_descriptive_stats.R
\name{plot_metabolite_histogram}
\alias{plot_metabolite_histogram}
\title{Visualize distribution of metabolite concentrations}
\usage{
plot_metabolite_histogram(data, metabolite_name)
}
\arguments{
\item{data}{A data frame containing metabolomics data.}

\item{metabolite_name}{Name of the metabolite to visualize.}
}
\value{
A ggplot object displaying histogram of metabolite concentrations.
}
\description{
This function creates histograms to visualize the distribution of metabolite concentrations.
}
\examples{
data <- MDAW::table_data
plot_metabolite_histogram(data, "M1")

}
