% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_to_excel.R
\name{save_to_xlsx}
\alias{save_to_xlsx}
\title{Save a tibble to an XLSX file}
\usage{
save_to_xlsx(data, path)
}
\arguments{
\item{data}{A tibble to be saved.}

\item{path}{The file path where the XLSX file will be saved.}
}
\value{
Nothing is returned. The tibble is saved to the specified path.
}
\description{
[`save_to_xlsx`] saves a tibble (data frame) to an XLSX file
using the writexl package.
}
\examples{
library(tibble)
library(MDAW)

data <- tibble(x = 1:5, y = letters[1:5])
data

# MDAW::save_to_xlsx(data, "data.xlsx")
}
