% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data_train_test.R
\name{split_data_train_test}
\alias{split_data_train_test}
\title{Split Data into Training and Test Sets}
\usage{
split_data_train_test(data, split_ratio, seed = NULL)
}
\arguments{
\item{data}{A data frame or matrix containing the dataset to be split.}

\item{split_ratio}{A numeric value between 0 and 1 indicating the proportion
of the dataset to include in the training set. The rest will be included
in the test set.}

\item{seed}{An optional seed for reproducibility.}
}
\value{
A list containing two elements: 'train' for the training set
and 'test' for the test set.
}
\description{
This function takes a dataset and splits it into training and test sets
based on a specified ratio.
}
\examples{
data(iris)
split_data <- split_data_train_test(iris, split_ratio = 0.8, seed = 123)
dim(split_data$train)
dim(split_data$test)

}
