% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_peak}
\alias{table_peak}
\title{Metabolite Peak with the Interactive dataTable}
\format{
\code{table_peak} This table encompasses 149 metabolites and features,
in a data frame with 129 rows and 4 columns:
\describe{
\item{RowID}{The \code{RowID} column represents row id of the metabolite}
\item{Idx}{The \code{idX} column represents index of the metabolite}
\item{Name}{The \code{Name} column represents name of the metabolite}
\item{Label}{The \code{Label} column represents chemical nomeclature
of the metabolite}
}
}
\source{
\url{https://www.metabolomicsworkbench.org/data/DRCCMetadata.php?Mode=Project&ProjectID=PR000699}
}
\usage{
table_peak
}
\description{
Each row within the table_peak corresponds to a METABOLITE,
with the following column descriptions:
}
\keyword{datasets}
