% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_2class.R
\name{univariate_2class}
\alias{univariate_2class}
\title{Univariate Statistics for Two-Class Data}
\usage{
univariate_2class(
  DataTable,
  PeakTable,
  group,
  posclass,
  parametric = TRUE,
  seed = NULL
)
}
\arguments{
\item{DataTable}{Data sheet with the required columns.}

\item{PeakTable}{Peak sheet with the required columns.}

\item{group}{Name of the column in the DataTable that contains
the Class data.}

\item{posclass}{Name of the positive class in the group column.}

\item{parametric}{If \code{TRUE}, parametric statistics (mean, t-test)
are calculated. If \code{FALSE}, non-parametric statistics
(median, Mann-Whitney U test) are calculated. i.e. Median 95\%CI
and Mann-Whitney U test is calculated instead of Mean 95\%CI and t-test.}

\item{seed}{Used to seed the generator for  the Median 95\% CI bootstrap
(resample with replacement) when calculating bootstrap confidence intervals.
Ignored if \code{parametric} is \code{TRUE}.
Note, if parametric=TRUE the seed is irrelevant as Mean 95\% CI
is calculated parametrically.}
}
\value{
A data frame containing univariate statistics for each peak.
The table  should contain multiple univariate statistics (2 class).
}
\description{
Calculates univariate statistics for two-class data including mean, median,
confidence intervals, and statistical tests such as t-test
or Mann-Whitney U test.
}
\examples{
# Example usage
univariate_2class(DataTable, PeakTable, group = "Class", posclass = "A")

}
