% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dataXL.R
\name{import_dataXL}
\alias{import_dataXL}
\title{Import Data and Peak tables from xslx}
\usage{
import_dataXL(xlsx_file_name, sheet_name, interactive_output = FALSE)
}
\arguments{
\item{xlsx_file_name}{A name of excel file.}

\item{sheet_name}{A name of indicating the sheet of interest.}

\item{interactive_output}{Boolean indicating whether to output
the table in interactive mode using reactable.}
}
\value{
The table of the \code{tibble} format or as a reactable
if interactive_output is \emph{TRUE}.
}
\description{
\code{\link[=import_dataXL]{import_dataXL()}} reads sheets from xslx and load them.
}
\examples{
library(MDAW)
excel_file <- "Gastric_NMR.xlsx"
path <- system.file("extdata", "Gastric_NMR.xlsx", package = "MDAW")
MDAW::import_dataXL(xlsx_file_name = path,
    sheet_name = "data")

}
