% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDAW-package.R
\docType{package}
\name{MDAW-package}
\alias{MDAW}
\alias{MDAW-package}
\title{MDAW: Metabolomic Data Analysis Workflow}
\description{
Enhancing Reproducibility: A Computational Workflow for Statistical Analysis of Untargeted and Targeted Metabolomics Data.
}
\author{
\strong{Maintainer}: Dany Mukesha \email{danymukesha@gmail.com} (\href{https://orcid.org/0009-0001-9514-751X}{ORCID})

}
\keyword{internal}
