% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDAW-package.R
\docType{package}
\name{MDAW-package}
\alias{MDAW}
\alias{MDAW-package}
\title{MDAW: Metabolomic Data Analysis Workflow}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Enhancing Reproducibility: A Computational Workflow for Statistical Analysis of Untargeted and Targeted Metabolomics Data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://danymukesha.github.io/MDAW/}
  \item \url{https://github.com/danymukesha/MDAW}
  \item Report bugs at \url{https://github.com/danymukesha/MDAW/issues}
}

}
\author{
\strong{Maintainer}: Dany Mukesha \email{danymukesha@gmail.com} (\href{https://orcid.org/0009-0001-9514-751X}{ORCID})

}
\keyword{internal}
